requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
psycopg2-binary
passlib[bcrypt]
python-jose
pydantic[email]
aiohttp
reportlab
